+++ lib/src/src.dart
@@ -1,0 +1,2 @@
+export 'apps/apps.dart';
+export 'core/core.dart';