+++ test/src/utils/launcher_util_test.dart
@@ -1,0 +1,2 @@
+// TODO: Add tests for the LauncherUtil class
+void main() {}